backend/requirements.txt
Django>=4.2
djangorestframework
pandas
openpyxl
django-cors-headers
gunicorn
python-dotenv
